Closes #58: - Moved all of the adapters in their own classes.
