+++ app/src/main/java/com/cesarsk/say_it/ui/fragments/HistoryFragment.java
@@ -1,39 +1,28 @@
 package com.cesarsk.say_it.ui.fragments;
 
 
-import android.app.ActivityOptions;
-import android.app.AlertDialog;
 import android.app.Fragment;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.Intent;
 import android.graphics.Canvas;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.media.AudioManager;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v4.content.ContextCompat;
-import android.support.v4.util.Pair;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
-import android.util.Log;
-import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.cesarsk.say_it.ui.MainActivity;
 import com.cesarsk.say_it.R;
-import com.cesarsk.say_it.ui.PlayActivity;
+import com.cesarsk.say_it.ui.components.HistoryAdapter;
 import com.cesarsk.say_it.utility.SayItPair;
 import com.cesarsk.say_it.utility.UtilitySharedPrefs;
 import com.google.gson.Gson;
@@ -43,8 +32,6 @@ import java.util.Collections;
 import java.util.Comparator;
 
 import uk.co.deanwild.materialshowcaseview.MaterialShowcaseView;
-
-import static android.speech.tts.TextToSpeech.QUEUE_FLUSH;
 
 
 /**
@@ -89,7 +76,7 @@ public class HistoryFragment extends Fragment {
         recyclerView.setLayoutManager(linearLayoutManager);
         recyclerView.setItemAnimator(defaultItemAnimator);
         recyclerView.addItemDecoration(dividerItemDecoration);
-        final HistoryAdapter adapter = new HistoryAdapter(deserializedHistory);
+        final HistoryAdapter adapter = new HistoryAdapter(this, deserializedHistory);
         recyclerView.setAdapter(adapter);
 
         ItemTouchHelper touchHelper = new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
@@ -245,154 +232,6 @@ public class HistoryFragment extends Fragment {
         return view;
     }
 
-    private class HistoryAdapter extends RecyclerView.Adapter<HistoryAdapter.ViewHolder> {
-
-        public static final long UNDO_TIMEOUT = 3000; //Timeout prima che l'elemento venga cancellato definitivamente
-
-        public ArrayList<SayItPair> getHistory() {
-            return history;
-        }
-
-        public void setHistory(ArrayList<SayItPair> history) {
-            this.history = history;
-            notifyDataSetChanged();
-        }
-
-        public SayItPair getTemp_hist() {
-            return temp_hist;
-        }
-
-        private SayItPair temp_hist;
-
-        private ArrayList<SayItPair> history;
-
-        public HistoryAdapter(ArrayList<SayItPair> history_list) {
-
-            history = history_list;
-        }
-
-        class ViewHolder extends RecyclerView.ViewHolder {
-
-            final TextView wordTextView;
-            final TextView IPATextView;
-            final ImageButton QuickPlayBtn;
-            final ImageButton AddtoFavsBtn;
-
-            ViewHolder(View itemView) {
-                super(itemView);
-                wordTextView = (TextView) itemView.findViewById(R.id.list_item_first_line);
-                IPATextView = (TextView) itemView.findViewById(R.id.list_item_second_line);
-                QuickPlayBtn = (ImageButton) itemView.findViewById(R.id.list_item_quickplay);
-                AddtoFavsBtn = (ImageButton) itemView.findViewById(R.id.list_item_addToFavs);
-            }
-        }
-
-        @Override
-        public HistoryAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-
-            LayoutInflater inflater = LayoutInflater.from(parent.getContext());
-
-            View view = inflater.inflate(R.layout.list_item_history, parent, false);
-
-            return new ViewHolder(view);
-        }
-
-        @Override
-        public void onBindViewHolder(final HistoryAdapter.ViewHolder holder, int position) {
-
-            holder.QuickPlayBtn.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (!isVolumeMuted()) {
-                        if (MainActivity.DEFAULT_ACCENT.equals("0")) {
-                            MainActivity.american_speaker_google.speak(holder.wordTextView.getText(), QUEUE_FLUSH, null, null);
-                            if(MainActivity.isLoggingEnabled) Log.i("DEFAULT - HISTORY", MainActivity.DEFAULT_ACCENT);
-                        } else if (MainActivity.DEFAULT_ACCENT.equals("1")) {
-                            MainActivity.british_speaker_google.speak(holder.wordTextView.getText(), QUEUE_FLUSH, null, null);
-                            if(MainActivity.isLoggingEnabled) Log.i("DEFAULT - HISTORY", MainActivity.DEFAULT_ACCENT);
-                        }
-                    } else {
-                        Toast toast = Toast.makeText(getActivity(), "Please turn the volume up", Toast.LENGTH_LONG);
-                        toast.setGravity(Gravity.CENTER, 0, 0);
-                        toast.show();
-                    }
-                }
-            });
-
-            holder.wordTextView.setText(history.get(position).first);
-            holder.IPATextView.setText(history.get(position).second);
-
-            holder.itemView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    final Intent play_activity_intent = new Intent(getActivity(), PlayActivity.class);
-                    play_activity_intent.putExtra(PlayActivity.PLAY_WORD, holder.wordTextView.getText());
-                    play_activity_intent.putExtra(PlayActivity.PLAY_IPA, holder.IPATextView.getText());
-                    UtilitySharedPrefs.addHist(getActivity(), new SayItPair(holder.wordTextView.getText().toString(), holder.IPATextView.getText().toString()));
-                    getActivity().startActivity(play_activity_intent, ActivityOptions.makeSceneTransitionAnimation(getActivity()).toBundle());
-                }
-            });
-
-            if (UtilitySharedPrefs.checkFavs(getActivity(), history.get(position).first))
-                holder.AddtoFavsBtn.setColorFilter(ContextCompat.getColor(getActivity(), R.color.RudolphsNose));
-
-            holder.AddtoFavsBtn.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (!UtilitySharedPrefs.checkFavs(getActivity(), history.get(holder.getAdapterPosition()).first)) {
-                        UtilitySharedPrefs.addFavs(getActivity(), new Pair<>(holder.wordTextView.getText().toString(), holder.IPATextView.getText().toString()));
-                        Toast.makeText(getActivity(), "Added to Favorites", Toast.LENGTH_SHORT).show();
-                        holder.AddtoFavsBtn.setColorFilter(ContextCompat.getColor(getActivity(), R.color.RudolphsNose));
-                    } else if (UtilitySharedPrefs.checkFavs(getActivity(), history.get(holder.getAdapterPosition()).first)) {
-                        UtilitySharedPrefs.removeFavs(v.getContext(), history.get(holder.getAdapterPosition()));
-                        Toast.makeText(getActivity(), "Removed from Favorites", Toast.LENGTH_SHORT).show();
-                        holder.AddtoFavsBtn.setColorFilter(ContextCompat.getColor(getActivity(), R.color.primary_dark));
-                    }
-                }
-            });
-
-            final FloatingActionButton fab = (FloatingActionButton) getActivity().findViewById(R.id.floating_button_history);
-            startTutorialPlayActivity(holder);
-            fab.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    new AlertDialog.Builder(getActivity())
-                            .setTitle("Clear History")
-                            .setMessage("Are you sure you want to clear your History?")
-                            .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int which) {
-                                    UtilitySharedPrefs.clearHistory(getActivity());
-                                    setHistory(loadDeserializedHistory(getActivity()));
-                                    Toast.makeText(getActivity(), "History Cleared!", Toast.LENGTH_SHORT).show();
-
-                                }
-                            })
-                            .setNegativeButton(android.R.string.no, new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int which) {
-                                    //do nothing
-                                }
-                            })
-                            .show();
-                }
-            });
-
-        }
-
-        @Override
-        public int getItemCount() {
-            return history.size();
-        }
-
-        void remove(int pos) {
-            temp_hist = history.get(pos);
-
-            UtilitySharedPrefs.removeHist(getActivity(), new SayItPair(history.get(pos).first, history.get(pos).second, history.get(pos).getAdding_time()));
-            history = loadDeserializedHistory(getActivity());
-            notifyItemRemoved(pos);
-        }
-
-    }
-
     public static ArrayList<SayItPair> loadDeserializedHistory(Context context) {
 
         UtilitySharedPrefs.loadHist(context);
@@ -415,7 +254,7 @@ public class HistoryFragment extends Fragment {
         return DeserializedHistory;
     }
 
-    private void startTutorialPlayActivity(HistoryAdapter.ViewHolder holder) {
+    public void startTutorialPlayActivity(HistoryAdapter.ViewHolder holder) {
         MainActivity.showCaseFragmentView = new MaterialShowcaseView.Builder(getActivity())
                 .setTarget(holder.wordTextView)
                 .setDismissText(getString(R.string.showcase_str_btn_5))
@@ -426,7 +265,7 @@ public class HistoryFragment extends Fragment {
                 .withoutShape()
                 .show();
     }
-    private boolean isVolumeMuted() {
+    public boolean isVolumeMuted() {
         int currentVolume = audio.getStreamVolume(AudioManager.STREAM_MUSIC);
         if (currentVolume == 0) return true;
         else return false;
